# src/main/gen/src/main/java/io/security/portfolio/api/DocumentPermissionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.security.portfolio.api;

import io.security.portfolio.dto.AddDocumentPermission201Response;
import io.security.portfolio.dto.DocumentPermissionRequest;
import io.security.portfolio.dto.ListDocumentPermissions200Response;
import io.security.portfolio.dto.ModelApiResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-07T12:21:57.339446635+09:00[Asia/Seoul]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "DocumentPermissions", description = "the DocumentPermissions API")
public interface DocumentPermissionsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/documents/{documentId}/permissions : 문서 권한 추가 (PERMIT/DENY)
     *
     * @param documentId  (required)
     * @param documentPermissionRequest  (required)
     * @return 생성된 문서 권한 정보 (status code 201)
     */
    @Operation(
        operationId = "addDocumentPermission",
        summary = "문서 권한 추가 (PERMIT/DENY)",
        tags = { "DocumentPermissions" },
        responses = {
            @ApiResponse(responseCode = "201", description = "생성된 문서 권한 정보", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AddDocumentPermission201Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/documents/{documentId}/permissions",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AddDocumentPermission201Response> addDocumentPermission(
        @Parameter(name = "documentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("documentId") Long documentId,
        @Parameter(name = "DocumentPermissionRequest", description = "", required = true) @Valid @RequestBody DocumentPermissionRequest documentPermissionRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"granteeId\" : 1, \"effect\" : \"effect\", \"action\" : \"action\", \"documentId\" : 6, \"id\" : 0, \"granteeType\" : \"granteeType\" }, \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/documents/{documentId}/permissions : 문서 권한 목록 조회
     *
     * @param documentId  (required)
     * @return 해당 문서에 대한 ACL 목록 (status code 200)
     */
    @Operation(
        operationId = "listDocumentPermissions",
        summary = "문서 권한 목록 조회",
        tags = { "DocumentPermissions" },
        responses = {
            @ApiResponse(responseCode = "200", description = "해당 문서에 대한 ACL 목록", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ListDocumentPermissions200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/documents/{documentId}/permissions",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ListDocumentPermissions200Response> listDocumentPermissions(
        @Parameter(name = "documentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("documentId") Long documentId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : [ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"granteeId\" : 1, \"effect\" : \"effect\", \"action\" : \"action\", \"documentId\" : 6, \"id\" : 0, \"granteeType\" : \"granteeType\" }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"granteeId\" : 1, \"effect\" : \"effect\", \"action\" : \"action\", \"documentId\" : 6, \"id\" : 0, \"granteeType\" : \"granteeType\" } ], \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /api/documents/{documentId}/permissions/{permissionId} : 문서 권한 제거
     *
     * @param documentId  (required)
     * @param permissionId  (required)
     * @return 삭제 결과 (status code 200)
     */
    @Operation(
        operationId = "removeDocumentPermission",
        summary = "문서 권한 제거",
        tags = { "DocumentPermissions" },
        responses = {
            @ApiResponse(responseCode = "200", description = "삭제 결과", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/documents/{documentId}/permissions/{permissionId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ModelApiResponse> removeDocumentPermission(
        @Parameter(name = "documentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("documentId") Long documentId,
        @Parameter(name = "permissionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("permissionId") Long permissionId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : \"\", \"success\" : true, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}