# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Document Authorization API
  version: 1.0.0
  description: |
    RBAC + Team Role + ABAC + ACL(PBAC 기반) 구조를 가진 문서 인가 데모 API.
    - 글로벌 RBAC: users / roles / user_roles
    - 팀/조직: teams / team_members
    - 문서: documents
    - ACL(문서별 권한): document_permissions
    - 감사 로그: audit_logs
    - 인증: refresh_tokens
servers:
  - url: http://localhost:8080
    description: Local dev server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        message:
          type: string
          nullable: true
          description: 추가 메시지 (없으면 null)
        data:
          nullable: true
          description: 실제 응답 데이터 (엔드포인트별로 타입 상이)

    # =====================
    # Auth
    # =====================
...
    DocumentPermissionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documentId:
          type: integer
          format: int64
        granteeType:
          type: string
          description: "USER, TEAM, ROLE"
        granteeId:
          type: integer
          format: int64
        action:
          type: string
          description: "예: document:read, document:write"
        effect:
          type: string
          description: "PERMIT 또는 DENY"
        createdAt:
          type: string
          format: date-time
    DocumentPermissionRequest:
      type: object
      required: [granteeType, granteeId, action, effect]
      properties:
        granteeType:
          type: string
          description: "USER, TEAM, ROLE"
        granteeId:
          type: integer
          format: int64
        action:
          type: string
        effect:
          type: string
          description: "PERMIT 또는 DENY"

    # =====================
    # Audit Log
    # =====================
...
paths:
...
  /api/documents/{documentId}/permissions:
    get:
      tags: [DocumentPermissions]
      summary: 문서 권한 목록 조회
      operationId: listDocumentPermissions
      security:
        - bearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 해당 문서에 대한 ACL 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/DocumentPermissionResponse'
    post:
      tags: [DocumentPermissions]
      summary: 문서 권한 추가 (PERMIT/DENY)
      operationId: addDocumentPermission
      security:
        - bearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentPermissionRequest'
      responses:
        '201':
          description: 생성된 문서 권한 정보
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DocumentPermissionResponse'
  /api/documents/{documentId}/permissions/{permissionId}:
    delete:
      tags: [DocumentPermissions]
      summary: 문서 권한 제거
      operationId: removeDocumentPermission
      security:
        - bearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: permissionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 삭제 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # =====================
  # Audit Logs
  # =====================
...